Move Leg and LegAdder in their own files
