+++ include/powsybl/iidm/ThreeWindingsTransformerAdder.hpp
@@ -10,6 +10,7 @@
 
 #include <powsybl/iidm/IdentifiableAdder.hpp>
 #include <powsybl/iidm/ThreeWindingsTransformer.hpp>
+#include <powsybl/iidm/ThreeWindingsTransformerLegAdder.hpp>
 #include <powsybl/stdcxx/math.hpp>
 
 namespace powsybl {
@@ -22,73 +23,7 @@ class VoltageLevel;
 
 class ThreeWindingsTransformerAdder : public IdentifiableAdder<ThreeWindingsTransformerAdder> {
 public:
-    class LegAdder : public virtual Validable {
-    public:  // Validable
-        std::string getMessageHeader() const override;
-
-    public:
-        ~LegAdder() noexcept override = default;
-
-        ThreeWindingsTransformerAdder& add();
-
-        LegAdder& setB(double b);
-
-        LegAdder& setBus(const std::string& bus);
-
-        LegAdder& setConnectableBus(const std::string& connectableBus);
-
-        LegAdder& setG(double g);
-
-        LegAdder& setNode(unsigned long node);
-
-        LegAdder& setR(double r);
-
-        LegAdder& setRatedS(double ratedS);
-
-        LegAdder& setRatedU(double ratedU);
-
-        LegAdder& setVoltageLevel(const std::string& voltageLevelId);
-
-        LegAdder& setX(double x);
-
-    private:
-        LegAdder(ThreeWindingsTransformerAdder& parent, unsigned long legNumber);
-
-        LegAdder& operator=(const LegAdder& adder);
-
-        ThreeWindingsTransformer::Leg build() const;
-
-        std::unique_ptr<Terminal> checkAndGetTerminal(VoltageLevel& voltageLevel);
-
-        VoltageLevel& checkAndGetVoltageLevel();
-
-        friend class ThreeWindingsTransformerAdder;
-
-    private:
-        ThreeWindingsTransformerAdder& m_parent;
-
-        unsigned long m_legNumber;
-
-        std::string m_voltageLevelId;
-
-        stdcxx::optional<unsigned long> m_node;
-
-        std::string m_bus;
-
-        std::string m_connectableBus;
-
-        double m_r = stdcxx::nan();
-
-        double m_x = stdcxx::nan();
-
-        double m_g = stdcxx::nan();
-
-        double m_b = stdcxx::nan();
-
-        double m_ratedU = stdcxx::nan();
-
-        double m_ratedS = stdcxx::nan();
-    };
+    using LegAdder = three_windings_transformer::LegAdder;
 
 public:
     explicit ThreeWindingsTransformerAdder(Substation& substation);
@@ -121,6 +56,8 @@ private:
     void setLegAdder2(LegAdder& legAdder);
 
     void setLegAdder3(LegAdder& legAdder);
+
+    friend class three_windings_transformer::LegAdder;
 
 private:
     Substation& m_substation;
