+++ src/iidm/ThreeWindingsTransformer.cpp
@@ -9,203 +9,12 @@
 
 #include <powsybl/iidm/Enum.hpp>
 #include <powsybl/iidm/PhaseTapChanger.hpp>
-#include <powsybl/iidm/PhaseTapChangerAdder.hpp>
 #include <powsybl/iidm/RatioTapChanger.hpp>
-#include <powsybl/iidm/RatioTapChangerAdder.hpp>
 #include <powsybl/iidm/Substation.hpp>
-#include <powsybl/iidm/ValidationUtils.hpp>
 
 namespace powsybl {
 
 namespace iidm {
-
-ThreeWindingsTransformer::Leg::Leg(unsigned long legNumber, double r, double x, double g, double b, double ratedU, double ratedS) :
-    m_legNumber(legNumber),
-    m_r(checkR(*this, r)),
-    m_x(checkX(*this, x)),
-    m_g(checkG(*this, g)),
-    m_b(checkB(*this, b)),
-    m_ratedU(checkRatedU(*this, ratedU, legNumber)),
-    m_ratedS(checkRatedS(*this, ratedS)) {
-}
-
-double ThreeWindingsTransformer::Leg::getB() const {
-    return m_b;
-}
-
-stdcxx::CReference<CurrentLimits> ThreeWindingsTransformer::Leg::getCurrentLimits() const {
-    return stdcxx::cref(m_limits);
-}
-
-stdcxx::Reference<CurrentLimits> ThreeWindingsTransformer::Leg::getCurrentLimits() {
-    return stdcxx::ref<CurrentLimits>(m_limits);
-}
-
-double ThreeWindingsTransformer::Leg::getG() const {
-    return m_g;
-}
-
-std::string ThreeWindingsTransformer::Leg::getMessageHeader() const {
-    return stdcxx::format("%1% '%2%': ", getTypeDescription(), m_transformer.get().getId());
-}
-
-const Network& ThreeWindingsTransformer::Leg::getNetwork() const {
-    return m_transformer.get().getNetwork();
-}
-
-Network& ThreeWindingsTransformer::Leg::getNetwork() {
-    return m_transformer.get().getNetwork();
-}
-
-const PhaseTapChanger& ThreeWindingsTransformer::Leg::getPhaseTapChanger() const {
-    if (!m_phaseTapChanger) {
-        throw PowsyblException("Phase tap changer not set");
-    }
-    return *m_phaseTapChanger;
-}
-
-PhaseTapChanger& ThreeWindingsTransformer::Leg::getPhaseTapChanger() {
-    if (!m_phaseTapChanger) {
-        throw PowsyblException("Phase tap changer not set");
-    }
-    return *m_phaseTapChanger;
-}
-
-double ThreeWindingsTransformer::Leg::getR() const {
-    return m_r;
-}
-
-double ThreeWindingsTransformer::Leg::getRatedS() const {
-    return m_ratedS;
-}
-
-double ThreeWindingsTransformer::Leg::getRatedU() const {
-    return m_ratedU;
-}
-
-const RatioTapChanger& ThreeWindingsTransformer::Leg::getRatioTapChanger() const {
-    if (!m_ratioTapChanger) {
-        throw PowsyblException("Ratio tap changer not set");
-    }
-    return *m_ratioTapChanger;
-}
-
-RatioTapChanger& ThreeWindingsTransformer::Leg::getRatioTapChanger() {
-    if (!m_ratioTapChanger) {
-        throw PowsyblException("Ratio tap changer not set");
-    }
-    return *m_ratioTapChanger;
-}
-
-unsigned long ThreeWindingsTransformer::Leg::getRegulatingTapChangerCount() const {
-    unsigned long count = 0;
-
-    const ThreeWindingsTransformer::Leg& leg1 = m_transformer.get().getLeg1();
-    count += (leg1.hasPhaseTapChanger() && leg1.getPhaseTapChanger().isRegulating()) ? 1 : 0;
-    count += (leg1.hasRatioTapChanger() && leg1.getRatioTapChanger().isRegulating()) ? 1 : 0;
-
-    const ThreeWindingsTransformer::Leg& leg2 = m_transformer.get().getLeg2();
-    count += (leg2.hasPhaseTapChanger() && leg2.getPhaseTapChanger().isRegulating()) ? 1 : 0;
-    count += (leg2.hasRatioTapChanger() && leg2.getRatioTapChanger().isRegulating()) ? 1 : 0;
-
-    const ThreeWindingsTransformer::Leg& leg3 = m_transformer.get().getLeg3();
-    count += (leg3.hasPhaseTapChanger() && leg3.getPhaseTapChanger().isRegulating()) ? 1 : 0;
-    count += (leg3.hasRatioTapChanger() && leg3.getRatioTapChanger().isRegulating()) ? 1 : 0;
-
-    return count;
-}
-
-const Terminal& ThreeWindingsTransformer::Leg::getTerminal() const {
-    return m_transformer.get().Connectable::getTerminal(m_legNumber - 1);
-}
-
-Terminal& ThreeWindingsTransformer::Leg::getTerminal() {
-    return m_transformer.get().Connectable::getTerminal(m_legNumber - 1);
-}
-
-const std::string& ThreeWindingsTransformer::Leg::getTypeDescription() const {
-    static std::array<std::string, 3> s_typeDescriptions {{
-        u8"3 windings transformer leg1",
-        u8"3 windings transformer leg2",
-        u8"3 windings transformer leg3"
-    }};
-
-    return s_typeDescriptions[m_legNumber - 1];
-}
-
-double ThreeWindingsTransformer::Leg::getX() const {
-    return m_x;
-}
-
-bool ThreeWindingsTransformer::Leg::hasPhaseTapChanger() const {
-    return static_cast<bool>(m_phaseTapChanger);
-}
-
-bool ThreeWindingsTransformer::Leg::hasRatioTapChanger() const {
-    return static_cast<bool>(m_ratioTapChanger);
-}
-
-CurrentLimitsAdder<const std::nullptr_t, ThreeWindingsTransformer::Leg> ThreeWindingsTransformer::Leg::newCurrentLimits() {
-    return CurrentLimitsAdder<const std::nullptr_t, Leg>(nullptr, *this);
-}
-
-ThreeWindingsTransformer::Leg& ThreeWindingsTransformer::Leg::setB(double b) {
-    m_b = checkB(*this, b);
-    return *this;
-}
-
-void ThreeWindingsTransformer::Leg::setCurrentLimits(std::nullptr_t /*side*/, std::unique_ptr<CurrentLimits> limits) {
-    m_limits = std::move(limits);
-}
-
-ThreeWindingsTransformer::Leg& ThreeWindingsTransformer::Leg::setG(double g) {
-    m_g = checkG(*this, g);
-    return *this;
-}
-
-ThreeWindingsTransformer::Leg& ThreeWindingsTransformer::Leg::setR(double r) {
-    m_r = checkR(*this, r);
-    return *this;
-}
-
-ThreeWindingsTransformer::Leg& ThreeWindingsTransformer::Leg::setRatedS(double ratedS) {
-    m_ratedS = checkRatedS(*this, ratedS);
-    return *this;
-}
-
-ThreeWindingsTransformer::Leg& ThreeWindingsTransformer::Leg::setRatedU(double ratedU) {
-    m_ratedU = checkRatedU(*this, ratedU);
-    return *this;
-}
-
-PhaseTapChangerAdder ThreeWindingsTransformer::Leg::newPhaseTapChanger() {
-    return PhaseTapChangerAdder(*this);
-}
-
-RatioTapChangerAdder ThreeWindingsTransformer::Leg::newRatioTapChanger() {
-    return RatioTapChangerAdder(*this);
-}
-
-ThreeWindingsTransformer::Leg& ThreeWindingsTransformer::Leg::setX(double x) {
-    m_x = checkX(*this, x);
-    return *this;
-}
-
-void ThreeWindingsTransformer::Leg::setPhaseTapChanger(std::unique_ptr<PhaseTapChanger> phaseTapChanger) {
-    m_phaseTapChanger = std::move(phaseTapChanger);
-}
-
-void ThreeWindingsTransformer::Leg::setRatioTapChanger(std::unique_ptr<RatioTapChanger> ratioTapChanger) {
-    m_ratioTapChanger = std::move(ratioTapChanger);
-}
-
-void ThreeWindingsTransformer::Leg::setTransformer(ThreeWindingsTransformer& transformer) {
-    m_transformer = stdcxx::ref(transformer);
-}
-
-std::string ThreeWindingsTransformer::Leg::toString() const {
-    return stdcxx::format("%1% leg%2%", m_transformer.get().getId(), m_legNumber);
-}
 
 ThreeWindingsTransformer::ThreeWindingsTransformer(const std::string& id, const std::string& name, bool fictitious, Leg&& leg1, Leg&& leg2, Leg&& leg3, double ratedU0) :
     Connectable(id, name, fictitious, ConnectableType::THREE_WINDINGS_TRANSFORMER),
